requirements.txt
aniso8601==2.0.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
enum34==1.1.6
Flask==0.12.2
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
future==0.16.0
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.6.0
python-dateutil==2.6.1
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
